Delete unused files.

+++ src/roughtime_parse.cpp
@@ -41,7 +41,7 @@ static const uint8_t * rp_subarray(
   sstring &out) {
   const size_t end = MIN(((int)toIdx), ((int)(bstr.length() - 1)));
   const size_t beg = MIN(end, fromIdx);
-  const int len = end - beg;
+  const int len = (int)end - (int)beg;
   LOG_ASSERT(len >= 0);
   out.clear();
   const uint8_t * const b = bstr.u_str();
@@ -122,7 +122,7 @@ uint64_t RoughTime::ParseToMicroseconds(
 
   int s = 0;
   int i = 0;
-  int n = b_length;
+  int n = (int)b_length;
 
   if (n % 4 > 0) {
     return rp_reject(b, "short message");
